Add dark mode colors
This overrides the colors when prefers-color-scheme is set to dark,
indicating the user has configured their OS or user-agent to prefer
a light text on dark background instead. It includes a way to manually
toggle dark mode by using a data attribute on the html element.

The syntax highlight colors are also updated to support dark mode.