Include all files in tarball
If you don't sepecify files: [...], then npm will use the .gitignore
which we don't want since it would ignore the dist folder
which we WANT included in the final .tar.gz package.

And, since we specify the files: [...] package.json entry, we must
explicitly list all files/folders that we want in there.